build.

+++ src/app/buyer/order/order.page.ts
@@ -42,32 +42,15 @@ export class OrderPage {
     this.router.navigateByUrl('buyer_order_detail', { state: order });
   }
 
-  async delete(order: Order) {
-    const alert = await utilsService.injector.get(AlertController).create({
-      //header: '确认!',
-      message: '确认删除此订单？',
-      buttons: [
-        {
-          text: '取消',
-          role: 'cancel',
-          cssClass: 'secondary',
-          handler: (blah) => {
-            console.log('Confirm Cancel: blah');
-          }
-        }, {
-          text: '确定',
-          handler: () => {
-            apiService.orderClient.delete(order, apiService.metaData, (err: any, response: any) => {
-              if (err) {
-                utilsService.alert(JSON.stringify(err));
-              } else {
-                this.ionViewWillEnter();
-              }
-            })
-          }
+  delete(order: Order) {
+    utilsService.confirm('确认删除此订单？', () => {
+      apiService.orderClient.delete(order, apiService.metaData, (err: any, response: any) => {
+        if (err) {
+          utilsService.alert(JSON.stringify(err));
+        } else {
+          this.ionViewWillEnter();
         }
-      ]
+      })
     });
-    await alert.present();
   }
 }
